chat-agent/pyproject.toml
[project]
name = "chat-agent"
version = "0.1.0"
description = "Case: Chat API com Strands Agents + Ollama"
authors = [
  { name = "Seu Nome", email = "you@example.com" }
]
requires-python = ">=3.10"

dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic>=2",
  "pydantic-settings",
  "strands-agents",
  "strands-agents-tools",
  "ollama",
]

[project.optional-dependencies]
dev = [
  "ruff",
  "mypy",
  "pytest",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
